+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 ### Added
 - Endpoint returning last transaction reference for a given address
+- Rewards metrics
+
+### Changed
+- Rewards adjusted to 4 snapshots per minute
 
 ## [v2.5.8] - 2020-05-01
 ### Fixed